fix decode numbers tests

+++ lib/src/test/kotlin/com/lukelast/ktoon/fixtures/decode/NumbersDecodeTest.kt
@@ -5,7 +5,8 @@ import kotlinx.serialization.Serializable
 import org.junit.jupiter.api.Test
 
 /**
- * Tests from numbers.json fixture - Number decoding edge cases: trailing zeros, exponent forms, negative zero.
+ * Tests from numbers.json fixture - Number decoding edge cases: trailing zeros, exponent forms,
+ * negative zero.
  */
 class NumbersDecodeTest {
 
@@ -13,71 +14,62 @@ class NumbersDecodeTest {
 
     @Test
     fun `parses number with trailing zeros in fractional part`() {
-        @Serializable
-        data class Value(val value: Double)
+        @Serializable data class Value(val value: Double)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses negative number with positive exponent`() {
-        @Serializable
-        data class Value(val value: Double)
+        @Serializable data class Value(val value: Int)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses lowercase exponent`() {
-        @Serializable
-        data class Value(val value: Double)
+        @Serializable data class Value(val value: Int)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses uppercase exponent with negative sign`() {
-        @Serializable
-        data class Value(val value: Double)
+        @Serializable data class Value(val value: Double)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses negative zero as zero`() {
-        @Serializable
-        data class Value(val value: Int)
+        @Serializable data class Value(val value: Int)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses negative zero with fractional part`() {
-        @Serializable
-        data class Value(val value: Int)
+        @Serializable data class Value(val value: Int)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses array with mixed numeric forms`() {
-        @Serializable
-        data class Nums(val nums: List<Double>)
+        // Note: all numbers are getting parsed to doubles.
+        @Serializable data class Nums(val nums: List<Double>)
         runDecodeFixtureTest<Nums>(fixture)
     }
 
     @Test
     fun `treats leading zero as string not number`() {
-        @Serializable
-        data class Value(val value: String)
+        @Serializable data class Value(val value: String)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses very small exponent`() {
-        @Serializable
-        data class Value(val value: Double)
+        @Serializable data class Value(val value: Double)
         runDecodeFixtureTest<Value>(fixture)
     }
 
     @Test
     fun `parses integer with positive exponent`() {
-        @Serializable
-        data class Value(val value: Int)
+        @Serializable data class Value(val value: Int)
         runDecodeFixtureTest<Value>(fixture)
     }
 
@@ -113,15 +105,13 @@ class NumbersDecodeTest {
 
     @Test
     fun `treats leading-zero in object value as string`() {
-        @Serializable
-        data class A(val a: String)
+        @Serializable data class A(val a: String)
         runDecodeFixtureTest<A>(fixture)
     }
 
     @Test
     fun `treats leading-zeros in array as strings`() {
-        @Serializable
-        data class Nums(val nums: List<String>)
+        @Serializable data class Nums(val nums: List<String>)
         runDecodeFixtureTest<Nums>(fixture)
     }
 }
